feat: update theme provider
