--- app/bootstrap.client.tsx
@@ -1,3 +1,0 @@
-import {handleDarkAndLightModeEls} from './utils/theme-provider.tsx'
-// fixup stuff before hydration
-handleDarkAndLightModeEls()
