+++ app/components/course-card.tsx
@@ -4,7 +4,7 @@ import {ArrowLink} from './arrow-button.tsx'
 import {ButtonLink} from './button.tsx'
 import {ArrowIcon} from './icons.tsx'
 import {H2, H3, Paragraph} from './typography.tsx'
-import {Themed} from '~/utils/theme-provider.tsx'
+import {Themed} from '~/utils/theme-utils.tsx'
 
 const MotionButtonLink = motion(ButtonLink)
