+++ app/entry.server.tsx
@@ -79,7 +79,9 @@ function serveTheBots(...args: DocRequestArgs) {
     remixContext,
     loadContext,
   ] = args
-  const nonce = loadContext.cspNonce ? String(loadContext.cspNonce) : undefined
+
+  const nonce = String(loadContext.cspNonce)
+
   return new Promise((resolve, reject) => {
     const stream = renderToPipeableStream(
       <NonceProvider value={nonce}>
@@ -132,7 +134,7 @@ function serveBrowsers(...args: DocRequestArgs) {
     remixContext,
     loadContext,
   ] = args
-  const nonce = loadContext.cspNonce ? String(loadContext.cspNonce) : undefined
+  const nonce = String(loadContext.cspNonce)
   return new Promise((resolve, reject) => {
     let didError = false
     const stream = renderToPipeableStream(
